Highlighting invalid fields and skip login class
Changes the colour of the label on an invalid field to red. It's fully
functional for all the person tabs.  Also added a copy of the main
driver that skips the login page. Should make things easier for you.
I've added some more comments as well and I'll keep working on this.