Fixed some error markers and added Jar file
Fixed incorrectly working error markers on some labels. Also included a
runnable JAR file of the entire project (just double click it in the
file browser to run the program - no need to open eclipse). We'll have
to create a new one when we're done, so this is just to demonstrate what
our finished program will do.